feature(api): changed api for decorators
Previously the worklet context decorator API was a bit misleading since you had to call it through the Worklet API for everything to be installed correctly.

With this change the default worklet context is a singleton that can be accessed with JsiWorkletContext::getInstance(). The list of decorators are also static and global for all worklet contexts.

Decorators are also decorated in the worklet context thread / runtime.

#fixes #19