fix: Added RuntimeAwareCache to make sure we release objects correctly
To avoid keeping references to shared_ptrs that are never released, we need to make sure they are destructed on the same runtime.

This is solved by using the RuntimeAwareCache model as in Skia to make sure these key objects are only released together with the runtime they belong in.

More information can be found in the comments of the WKTRuntimeAwareCache.h file.

WKTJsiWorklet.h:
Removed the runtime teardown code from the WorkletInvoker's destructor. This is now taken care of by using the RuntimeAwareCache

WKTJsiWorkletApi.h:
Removed the passing of the default WorkletContext when creating new JsiSharedValues - the thing we want is the runtime and we can get it ourselves inside the shared value ctor.

WKTJsiWorkletContext.cpp:
Added setting the main JS Runtime on the RuntimeAwareCache so that its logic is correctly performed.

WKTJsiHostObject.h/cpp:
- Changed using RuntimeAwareCache for JSI object host functions

Podfile:
- Added fix with _LIBCPP_ENABLE_CXX17_REMOVED_UNARY_BINARY_FUNCTION error in XCode

Added WKTRuntimeAwareCache/WKTRuntimeAwareLIfecycleMonitor